change reset time from 1s to 10

+++ files/usr/lib/reset-button-watch.sh
@@ -8,6 +8,8 @@ PORT=0xa00      # I/O port address
 MASK=0x04       # Bit 2 = 0 means pressed, 1 means not pressed
 INTERVAL=1      # polling interval (seconds)
 
+COUNTER=0
+
 while :; do
     byte=$(dd if=/dev/port bs=1 skip=$((PORT)) count=1 2>/dev/null \
            | hexdump -v -e '1/1 "%u"')
@@ -15,10 +17,17 @@ while :; do
     [ -z "$byte" ] && { echo "read error"; exit 1; }
 
     if [ $((byte & MASK)) -eq 0 ]; then
-        # Button is pressed - performing factory reset
-        firstboot -y
-        reboot
-        exit 0
+        # Button is pressed - increment counter
+        COUNTER=$((COUNTER + 1))
+        
+        if [ "$COUNTER" -eq 10 ]; then
+            # Button has been pressed for 10s in a row, perform factory reset
+            firstboot -y
+            reboot
+            exit 0
+        fi
+    else
+        COUNTER=0
     fi
 
     sleep "$INTERVAL"
